moved geospatial resolution test helper functions from helper to geospatial resolution test file. created a utils file, currently just for plotting helper functions
